requirements.txt
SQLAlchemy==2.0.27
fastapi==0.109.2
pydantic==2.6.2
uvicorn==0.27.1
python-dotenv==1.0.1
psycopg2==2.9.9